requirements-ok.txt
fastapi==0.113.0
fastapi-cli==0.0.7
uvicorn>=0.30.1
pandas==2.3.0
python-dotenv==1.1.0
python-multipart==0.0.20
sqlalchemy==2.0.41
tabulate==0.9.0

pydantic==2.7.4 
langchain-community==0.2.19
langchain-core==0.2.43
langchain-experimental==0.0.65
langchain-text-splitters==0.2.4
langchain~=0.2.0
langchain-google-genai==1.0.10
google-generativeai==0.7.2


